Updates
1. I changed the names of the classes to match what we hve for the rest (I amended the code accordingly too)

2. Instead of RoomBill, I changed that class to GrandTotalTabulator. As for room bill, I handled it in room_rate package already. This grand total tabulator adds room bill and room service bill

3. There are actually two different guest list manipulators. The GuestListManipylator is for the hotel class. That one manipulates the main guest list stored in hotel object. For reservation, I created a ReservationGuestListManipulator.

4. I also added the code that checks for valid input. Eg Year is greater than 2021 (Except for time)

I also added an attribute called reservationDetails to Guest already. So now the createReserver and createGuest function will accept a reservation object.

I have made it such that the ReservationGuestManipulator, which creates the guests and adds them to the reservation object’s guest list, passes in its reservation attribute to the createReserver and createGuest function when they are called. So everything is fine.

Ohh yea for createReservation, I also added Room.setReservation(reservation) to assign reservation object to the room
5.